requirements.txt
numpy==1.24.4
setuptools==60.2.0
git+https://github.com/samuelstanton/upcycle.git@master
python-levenshtein==0.12.2
wandb==0.10.28
biopython==1.79
tabulate==0.8.9
pypdb==2.0
aiohttp==3.7.4
aioredis==1.3.0
omegaconf==2.1.1
hydra-core==1.1.0
cloudpickle==1.6.0
pickle5==0.0.11
pymoo==0.6.1
tqdm==4.62.3
pandas==1.3.0
deepchem==2.6.0.dev20211207173323
transformers==4.6
networkx==2.6.3
selfies==2.0.0
cachetools==5.0.0
jupyter==1.0.0
statsmodels==0.13.2
seaborn==0.11.2
vina==1.2.3
pyscreener==1.1.1
polyleven==0.8
google-api-core==2.16.1